pyproject.toml
[tool.poetry]
    name        = "bell-avr-vmc-fusion-module"
    version     = "0.1.0"
    description = "Sensor fusion module for Bell AVR"
    license     = "MIT"
    readme      = "README.md"
    homepage    = "https://github.com/bellflight/AVR-VMC-Fusion-Module"
    authors     = ["Nathan Vaughn <nvaughn51@gmail.com>"]

[tool.poetry.dependencies]
    python             = ">=3.9,<3.12"
    loguru             = "^0.6.0"
    bell-avr-libraries = "0.2.0a0"
    numpy              = "^1.23.5"
    pymap3d            = "^2.9.1"

[tool.poetry.group.dev.dependencies]
    pre-commit  = "^3.1.1"
    pytest      = "^7.2.1"
    pytest-cov  = "^4.0.0"
    pytest-mock = "^3.10.0"

[tool.pyright]
    typeCheckingMode = "basic"
    venvPath         = "."
    venv             = ".venv"

[tool.ruff]
    ignore = ["E501"]
